Move Round Robin logic to a separate client
